bugok

+++ sources/ft_valeur_absolue.c
@@ -6,16 +6,16 @@
 /*   By: sait-ben <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/02/06 14:29:16 by sait-ben          #+#    #+#             */
-/*   Updated: 2017/02/06 14:30:40 by sait-ben         ###   ########.fr       */
+/*   Updated: 2017/02/17 12:30:06 by sait-ben         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "printf.h"
 
-unsigned int	ft_valeurabsolue(int nb)
+uintmax_t	ft_valeurabsolue(intmax_t nb)
 {
 	if (nb >= 0)
-		return ((unsigned int)nb);
+		return ((uintmax_t)nb);
 	nb = -nb;
-	return ((unsigned int)nb);
+	return ((uintmax_t)nb);
 }
